requirements.txt
streamlit
openai
azure-identity
python-dotenv
pymongo